Ikkoku: split and adjust to adapt SubtitleParse
