+++ Ikkoku/SubtileProcess/CJKpp.cs
@@ -1,11 +1,11 @@
 using System.Text;
-using Mobsub.AssTypes;
+using Mobsub.SubtitleParse.AssTypes;
 using Mobsub.SubtitleParse;
 using OpenCCSharp.Conversion;
 
-namespace Mobsub.Ikkoku;
+namespace Mobsub.Ikkoku.SubtileProcess;
 
-public partial class SubtileProcess
+public class CJKpp
 {
     public static void ZhConvertEventLineByOpenccsharp(List<char[]> etText, StringBuilder sb, ChainedScriptConverter converter, out string[]? charsCountChange)
     {
