+++ Ikkoku/SubtileProcess/Check.cs
@@ -1,10 +1,10 @@
 using System.Text;
-using Mobsub.AssTypes;
+using Mobsub.SubtitleParse.AssTypes;
 using Mobsub.SubtitleParse;
 
-namespace Mobsub.Ikkoku;
+namespace Mobsub.Ikkoku.SubtileProcess;
 
-public partial class SubtileProcess
+public class Check
 {
     public static void PrintUnnormalAssTags(List<AssEvent> events, bool verbose, string checkType)
     {
@@ -197,16 +197,16 @@ public partial class SubtileProcess
     }
 
     public static bool WeridTimeOneLine(AssEvent evt) => evt.IsDialogue && (evt.Start.CompareTo(evt.End) > 0);
-    
+
     /// <summary>
     /// such like {=} {=0} {=99}
     /// </summary>
     /// <param name="et">Event.Text</param>
     /// <returns></returns>
-    private static bool IsMotionGarbage(List<char[]> et) => et.Count > 0 && AssTagParse.IsOvrrideBlock(et[0].AsSpan()) && et[0][1] == '=' && ((et[0].Length > 3 && char.IsDigit(et[0][2])) || et[0].Length == 3);
+    public static bool HadMotionGarbage(List<char[]> et) => et.Count > 0 && AssTagParse.IsOvrrideBlock(et[0].AsSpan()) && et[0][1] == '=' && ((et[0].Length > 3 && char.IsDigit(et[0][2])) || et[0].Length == 3);
 
-    private static readonly char[] EventUnusedChars = ['\u200E', '\u200F', '\u200B'];
-    private static readonly char[] EventWeirdSpace = ['\u00A0'];
+    public static readonly char[] EventUnusedChars = ['\u200E', '\u200F', '\u200B'];
+    public static readonly char[] EventWeirdSpace = ['\u00A0'];
 
     public static void CheckWeridChars(List<char[]> et, out bool hadUnusedChar, out bool hadWeridSpace)
     {
